Bug fixes
Fixed gallery videos from different apps not working

Updated workflow to start on welcome and go onto home to prevent endless loop on first start with some granted permissions
